# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Project API Course Item",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/Teachers": {
            "get": {
                "description": "list Teacher entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List Teacher entities",
                "operationId": "list-Teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Teacher"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Teacher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Teacher",
                "operationId": "create-Teacher",
                "parameters": [
                    {
                        "description": "Teacher entity",
                        "name": "Teacher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Teacher"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/Teachers/{id}": {
            "get": {
                "description": "get Teacher by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a Teacher entity by ID",
                "operationId": "get-Teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Teacher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Teacher"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update Teacher by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a Teacher entity by ID",
                "operationId": "update-Teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Teacher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Teacher entity",
                        "name": "Teacher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Teacher"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Teacher"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get Teacher by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a Teacher entity by ID",
                "operationId": "delete-Teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Teacher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ent.Course": {
            "type": "object",
            "properties": {
                "course_name": {
                    "description": "CourseName holds the value of the \"course_name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CourseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.CourseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.CourseEdges": {
            "type": "object",
            "properties": {
                "courseItems": {
                    "description": "CourseItems holds the value of the course_items edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CourseItem"
                    }
                },
                "owner": {
                    "description": "Owner holds the value of the owner edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Teacher"
                }
            }
        },
        "ent.CourseItem": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CourseItemQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.CourseItemEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.CourseItemEdges": {
            "type": "object",
            "properties": {
                "courses": {
                    "description": "Courses holds the value of the courses edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Course"
                },
                "subjects": {
                    "description": "Subjects holds the value of the subjects edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Subject"
                },
                "types": {
                    "description": "Types holds the value of the types edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SubjectType"
                }
            }
        },
        "ent.Subject": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SubjectQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SubjectEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "subject_name": {
                    "description": "SubjectName holds the value of the \"subject_name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.SubjectEdges": {
            "type": "object",
            "properties": {
                "courseItems": {
                    "description": "CourseItems holds the value of the course_items edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CourseItem"
                    }
                },
                "owner": {
                    "description": "Owner holds the value of the owner edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Teacher"
                }
            }
        },
        "ent.SubjectType": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SubjectTypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SubjectTypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "type_name": {
                    "description": "TypeName holds the value of the \"type_name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.SubjectTypeEdges": {
            "type": "object",
            "properties": {
                "courseItems": {
                    "description": "CourseItems holds the value of the course_items edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CourseItem"
                    }
                }
            }
        },
        "ent.Teacher": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TeacherQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TeacherEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "teacher_email": {
                    "description": "TeacherEmail holds the value of the \"teacher_email\" field.",
                    "type": "string"
                },
                "teacher_name": {
                    "description": "TeacherName holds the value of the \"teacher_name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.TeacherEdges": {
            "type": "object",
            "properties": {
                "courses": {
                    "description": "Courses holds the value of the courses edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Course"
                    }
                },
                "subjects": {
                    "description": "Subjects holds the value of the subjects edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Subject"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}